Chore: url 경로 재지정 및 컨트롤러 정리
