+++ src/main/java/com/matdang/seatdang/application/controller/MainController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class MainController {
 
-    @GetMapping("/")
+    @GetMapping("/my-seat-dang")
     public String customerMain(){
         return "customer/main";
     }
