+++ src/main/resources/application-test.yml
@@ -22,7 +22,7 @@ spring:
         format_sql: true
         #        show_sql: true
         hbm2ddl:
-          auto: create
+          auto: update
         dialect: org.hibernate.dialect.MySQLDialect
     open-in-view: false
 kakao:
