+++ src/main/resources/templates/customer/main.html
@@ -16,7 +16,7 @@
                     <div class="col-xl-9 col-lg-10 col-md-8">
                         <h1>마 싯 당</h1>
                         <p>우리는 빵에 살고 빵에 죽는다</p>
-                        <form method="get" th:action="@{/search}">
+                        <form method="get" th:action="@{/my-seat-dang/search}">
                             <div class="row g-0 custom-search-input">
                                 <div class="col-lg-4">
                                     <div class="form-group">
