+++ src/main/resources/templates/customer/search/search.html
@@ -29,7 +29,7 @@
 <!--                        검색 -->
                     </div>
                     <div class="col-xl-4 col-lg-5 col-md-5">
-                        <form th:action="@{/search}" method="get" >
+                        <form th:action="@{/my-seat-dang/search}" method="get" >
                             <div class="search-wrapper">
                                 <div class="input-group">
                                     <div class="input-field">
@@ -117,7 +117,7 @@
                                 <figure>
                                     <span class="ribbon off">웨이팅N분</span>
                                     <img th:src="@{/img/customer/detail_1.jpg}" th:data-src="${store.getThumbnail()}" class="img-fluid lazy" alt="">
-                                    <a th:href="|@{/customer/store/storeDetail/}${store.storeId}|" class="strip_info">
+                                    <a th:href="|@{/my-seat-dang/store/detail/}${store.storeId}|" class="strip_info">
                                         <small>케이크</small>
                                         <div class="item_title">
                                             <h3 th:text="${store.storeName}"></h3>
@@ -192,7 +192,7 @@
                             content: `
                                         <div class="infoBox" style="position: absolute; visibility: visible; width: 290px; left: 25px; bottom: -1px;">
                                         <div class="marker_info" id="marker_info">
-                                        <img src="https://kr.object.ncloudstorage.com/myseatdang-bucket/sample-folder/af4bc57e-5faa-493f-a8ea-2cf9bf2c6f13.png" alt="" style="width: 240px; height: 140px; display: block; margin: 0 auto; cursor: pointer;" onclick="window.location.href='/customer/store/storeDetail/${store.storeId}'">
+                                        <img src="https://kr.object.ncloudstorage.com/myseatdang-bucket/sample-folder/af4bc57e-5faa-493f-a8ea-2cf9bf2c6f13.png" alt="" style="width: 240px; height: 140px; display: block; margin: 0 auto; cursor: pointer;" onclick="window.location.href='/my-seat-dang/store/detail/${store.storeId}'">
                                         <span>
                                             <span class="infobox_rate">평점 | 7.5</span><h3>${store.storeName}</h3><em>제과점</em><strong>undefined</strong>
                                             <a class="btn_infobox_detail"></a>
